+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v4.19.0
 
 ### v4.19.0 Future Deprecation
+
 - `[CSS]` The Soho light theme CSS file has been renamed from `light-theme.css` to `theme-soho-light.css` ([1972](https://github.com/infor-design/enterprise/issues/1972))
 - `[CSS]` The Soho dark theme CSS file has been renamed from `dark-theme.css` to `theme-soho-dark.css` ([1972](https://github.com/infor-design/enterprise/issues/1972))
 - `[CSS]` The Soho high-contrast theme CSS file has been renamed from `high-contrast-theme.css` to `theme-soho-high-contrast.css` ([1972](https://github.com/infor-design/enterprise/issues/1972))